✨ (src/git): Add function to get commit changes
This commit adds a new function `get_commit_changes` to the `git.rs` file. This function uses the `git diff --cached` command to retrieve the changes in the working directory. It returns an `Option<Vec<String>>` where `None` is returned if there are no changes, and `Some(diff)` is returned if there are changes. The function skips the first line of the diff output and collects the remaining lines into a vector of strings.